[saihelper]: Create initSaiRedis function and put related code inside it
Move SAI Redis initialization related code outside main function.
In the future, global variables could be eliminated to make code cleaner.
getTimestamp could be moved to common library as well.
