+++ orchagent/saihelper.h
@@ -1,3 +1,6 @@
 #pragma once
 
+#include <string>
+
 void initSaiApi();
+void initSaiRedis(const std::string &record_location);
